requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
reportlab==4.0.7
clearbit==0.1.7
newspaper3k==0.2.8
pandas==2.1.4
lxml==4.9.3
googlenews==1.6.8